tests/wpt/web-platform-tests/css/css-tables/height-distribution/computing-row-measure-0.html

<!doctype html>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<link rel='stylesheet' href='../support/base.css' />

<link rel="author" title="Greg Whitworth" href="gwhit@microsoft.com" />
<link rel="help" href="https://drafts.csswg.org/css-tables-3/#computing-column-measures" />
<main>
    <h4>"Computing row measures"</h4>
    <p>This is testing that the table root's minimum is max(table-root width, capmin, gridmin) <a href="https://drafts.csswg.org/css-tables-3/#computing-the-table-width">Spec Text</a></p>

    <hr/>
    <table id="one" cellspacing="0" cellpadding="0">
        <tbody style="height: 1px;">
            <tr style="height: 10px;">
                <td style="height: 1px;"></td>
            </tr>
        </tbody>
    </table>

    <table id="two" cellspacing="0" cellpadding="0">
        <tbody style="height: 10px;">
            <tr style="height: 1px;">
                <td style="height: 1px;"></td>
            </tr>
        </tbody>
    </table>

    <table id="three" cellspacing="0" cellpadding="0">
        <tbody style="height: 1px;">
            <tr style="height: 1px;">
                <td style="height: 10px;"></td>
            </tr>
        </tbody>
    </table>

    <table id="four" cellspacing="0" cellpadding="0">
        <tbody style="height: 10px;">
            <tr>
                <td style="height: 1px;"></td>
            </tr>
        </tbody>
    </table>
</main>

<script>
    var i = 1;
    generate_tests(assert_equals, [
        [
            "Checking intermediate min-content height for span 1 (1)",
            document.getElementById('one').offsetHeight,
            10
        ],
        [
            "Checking intermediate min-content height for span 1 (2)",
            document.getElementById('two').offsetHeight,
            1
        ],
        [
            "Checking intermediate min-content height for span 1 (3)",
            document.getElementById('three').offsetHeight,
            10
        ],
        [
            "Checking intermediate min-content height for span 1 (4)",
            document.getElementById('four').offsetHeight,
            10
        ]
    ]);
</script>
</html>
