LayoutTests/imported/w3c/web-platform-tests/css/css-tables/height-distribution/computing-row-measure-0-expected.txt

"Computing row measures"

This is testing that the table root's minimum is max(table-root width, capmin, gridmin) Spec Text


PASS Checking intermediate min-content height for span 1 (1)
FAIL Checking intermediate min-content height for span 1 (2) assert_equals: expected 10 but got 1
PASS Checking intermediate min-content height for span 1 (3)
FAIL Checking intermediate min-content height for span 1 (4) assert_equals: expected 10 but got 1
